Lower Layer Super Output Area 2001 to 2011 lookup

A best fit lookup between lower layer super output areas as at December 2001 and lower layer super output areas and local authority districts as at December 2011 in England and Wales.

Usage

lookup_lsoa01_lsoa11

Format

A data frame of class "tbl" with 34,917 rows and 5 variables:

lsoa01_name

LSOA 2001 name

lsoa01_code

LSOA 2001 code

change_code

U = no change; S = split; M = merged; X = an irregular and fragmented relationship

lsoa11_name

LSOA 2011 name

lsoa11_code

LSOA 2011 code